+++ anknotes/EvernoteNotes.py
@@ -1,14 +1,15 @@
 # -*- coding: utf-8 -*-
-import os, time, re
+### Python Imports
 import copy
-from enums import AutoNumber, EvernoteTitleLevels
-# import EvernoteNotes as EN 
-from shared import *
-from operator import itemgetter 
+from operator import itemgetter
 try:
     from pysqlite2 import dbapi2 as sqlite
 except ImportError:
     from sqlite3 import dbapi2 as sqlite
+
+### Anknotes Imports
+# from anknotes.shared import *
+from anknotes.toc import *
      
 class NoteProcessingFlags:
     delayProcessing=False
