+++ anknotes/oauth2/__init__.py
@@ -29,6 +29,7 @@ import random
 import urlparse
 import hmac
 import binascii
+
 import httplib2
 
 try:
