+++ anknotes/version.py
@@ -26,8 +26,10 @@ Every version number class implements the following interface:
     of the same class, thus must follow the same rules)
 """
 
-import string, re
+import string
+import re
 from types import StringType
+
 
 class Version:
     """Abstract base class for version numbering classes.  Just provides
